+++ demo/src/test/java/fr/speekha/httpmocker/demo/model/DatasourceTest.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2019 David Blanc
+ * Copyright 2019-2020 David Blanc
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,11 +16,11 @@
 
 package fr.speekha.httpmocker.demo.model
 
-import fr.speekha.httpmocker.MockResponseInterceptor
 import fr.speekha.httpmocker.Mode
-import fr.speekha.httpmocker.builder.mockInterceptor
 import fr.speekha.httpmocker.demo.service.GithubApiEndpoints
 import fr.speekha.httpmocker.jackson.JacksonMapper
+import fr.speekha.httpmocker.okhttp.MockResponseInterceptor
+import fr.speekha.httpmocker.okhttp.builder.mockInterceptor
 import fr.speekha.httpmocker.policies.SingleFolderPolicy
 import kotlinx.coroutines.runBlocking
 import okhttp3.OkHttpClient
