+++ mocker-ktor/src/main/kotlin/fr/speekha/httpmocker/ktor/builder/MockableClientConfiguration.kt
@@ -14,9 +14,10 @@
  * limitations under the License.
  */
 
-package fr.speekha.httpmocker.ktor.engine
+package fr.speekha.httpmocker.ktor.builder
 
 import fr.speekha.httpmocker.builder.RecorderBuilder
+import fr.speekha.httpmocker.ktor.engine.MockEngineConfig
 import io.ktor.client.HttpClient
 import io.ktor.client.HttpClientConfig
 import io.ktor.client.engine.HttpClientEngineConfig
