+++ mocker-okhttp/src/main/kotlin/fr/speekha/httpmocker/okhttp/builder/InterceptorBuilder.kt
@@ -14,11 +14,13 @@
  * limitations under the License.
  */
 
-package fr.speekha.httpmocker.builder
+package fr.speekha.httpmocker.okhttp.builder
 
-import fr.speekha.httpmocker.MockResponseInterceptor
 import fr.speekha.httpmocker.Mode
 import fr.speekha.httpmocker.NO_ROOT_FOLDER_ERROR
+import fr.speekha.httpmocker.builder.ConfigBuilder
+import fr.speekha.httpmocker.builder.FileLoader
+import fr.speekha.httpmocker.okhttp.MockResponseInterceptor
 import fr.speekha.httpmocker.policies.FilingPolicy
 import fr.speekha.httpmocker.scenario.RequestCallback
 import fr.speekha.httpmocker.serialization.Mapper
